requirements.txt
prometheus-client 
psycopg2-binary 
pymongo 
boto3 
google-cloud-storage 
azure-storage-blob